reworked broken scoping rules for variable lookup
Before, variables with constant values were stored
in the global root environment, while variables with
address values were stored in their local scopes.
This was utterly broken and led to very strange
and unintended behaviour. Now we use Either to
store both types of values in the same Data.Map,
and this is done in the inner scope of the environment.

I.e. the semantics are a lot saner than before!